// packages/janribkaui-material-ui-tailwind/src/styles/props.d.ts
export interface ComponentsPropsList {
  // MuiAlert: AlertProps;
  // MuiAlertTitle: AlertTitleProps;
  // MuiAppBar: AppBarProps;
  // MuiAutocomplete: AutocompleteProps<any, any, any, any>;
  // MuiAvatar: AvatarProps;
  // MuiAvatarGroup: AvatarGroupProps;
  // MuiBackdrop: BackdropProps;
  // MuiBadge: BadgeProps;
  // MuiBottomNavigation: BottomNavigationProps;
  // MuiBottomNavigationAction: BottomNavigationActionProps;
  // MuiBreadcrumbs: BreadcrumbsProps;
  JrButton: ButtonProps;
  JrButtonBase: ButtonBaseProps;
  // MuiButtonGroup: ButtonGroupProps;
  // MuiCard: CardProps;
  // MuiCardActionArea: CardActionAreaProps;
  // MuiCardActions: CardActionsProps;
  // MuiCardContent: CardContentProps;
  // MuiCardHeader: CardHeaderProps;
  // MuiCardMedia: CardMediaProps;
  JrCheckbox: CheckboxProps;
  // MuiChip: ChipProps;
  JrCircularProgress: CircularProgressProps;
  // MuiCollapse: CollapseProps;
  // MuiContainer: ContainerProps;
  // MuiCssBaseline: CssBaselineProps;
  // MuiDialog: DialogProps;
  // MuiDialogActions: DialogActionsProps;
  // MuiDialogContent: DialogContentProps;
  // MuiDialogContentText: DialogContentTextProps;
  // MuiDialogTitle: DialogTitleProps;
  // MuiDivider: DividerProps;
  // MuiDrawer: DrawerProps;
  // MuiAccordion: AccordionProps;
  // MuiAccordionActions: AccordionActionsProps;
  // MuiAccordionDetails: AccordionDetailsProps;
  // MuiAccordionSummary: AccordionSummaryProps;
  // MuiFab: FabProps;
  // MuiFilledInput: FilledInputProps;
  JrFormControl: FormControlProps;
  JrFormControlLabel: FormControlLabelProps;
  JrFormGroup: FormGroupProps;
  JrFormHelperText: FormHelperTextProps;
  JrFormLabel: FormLabelProps;
  // MuiGrid: GridProps;
  // MuiGrid2: Grid2Props;
  // MuiImageList: ImageListProps;
  // MuiImageListItem: ImageListItemProps;
  // MuiImageListItemBar: ImageListItemBarProps;
  // MuiIcon: IconProps;
  JrIconButton: IconButtonProps;
  // MuiInput: InputProps;
  // MuiInputAdornment: InputAdornmentProps;
  // MuiInputBase: InputBaseProps;
  // MuiInputLabel: InputLabelProps;
  JrLinearProgress: LinearProgressProps;
  // MuiLink: LinkProps;
  // MuiList: ListProps;
  // MuiListItem: ListItemProps;
  // MuiListItemButton: ListItemButtonProps;
  // MuiListItemAvatar: ListItemAvatarProps;
  // MuiListItemIcon: ListItemIconProps;
  // MuiListItemSecondaryAction: ListItemSecondaryActionProps;
  // MuiListItemText: ListItemTextProps;
  // MuiListSubheader: ListSubheaderProps;
  // MuiMenu: MenuProps;
  // MuiMenuItem: MenuItemProps;
  // MuiMenuList: MenuListProps;
  // MuiMobileStepper: MobileStepperProps;
  // MuiModal: ModalProps;
  // MuiNativeSelect: NativeSelectProps;
  // MuiOutlinedInput: OutlinedInputProps;
  // MuiPagination: PaginationProps;
  // MuiPaginationItem: PaginationItemProps;
  // MuiPaper: PaperProps;
  // MuiPopper: PopperProps;
  // MuiPopover: PopoverProps;
  // MuiRadio: RadioProps;
  // MuiRadioGroup: RadioGroupProps;
  // MuiRating: RatingProps;
  // MuiScopedCssBaseline: ScopedCssBaselineProps;
  // MuiSelect: SelectProps;
  // MuiSkeleton: SkeletonProps;
  // MuiSlider: SliderProps;
  // MuiSnackbar: SnackbarProps;
  // MuiSnackbarContent: SnackbarContentProps;
  // MuiSpeedDial: SpeedDialProps;
  // MuiSpeedDialAction: SpeedDialActionProps;
  // MuiSpeedDialIcon: SpeedDialIconProps;
  // MuiStack: StackProps;
  // MuiStep: StepProps;
  // MuiStepButton: StepButtonProps;
  // MuiStepConnector: StepConnectorProps;
  // MuiStepContent: StepContentProps;
  // MuiStepIcon: StepIconProps;
  // MuiStepLabel: StepLabelProps;
  // MuiStepper: StepperProps;
  // MuiSvgIcon: SvgIconProps;
  // MuiSwipeableDrawer: SwipeableDrawerProps;
  JrSwitch: SwitchProps;
  // MuiTab: TabProps;
  // MuiTable: TableProps;
  // MuiTableBody: TableBodyProps;
  // MuiTableCell: TableCellProps;
  // MuiTableContainer: TableContainerProps;
  // MuiTableFooter: TableFooterProps;
  // MuiTableHead: TableHeadProps;
  // MuiTablePagination: TablePaginationProps;
  // MuiTableRow: TableRowProps;
  // MuiTableSortLabel: TableSortLabelProps;
  // MuiTabs: TabsProps;
  // MuiTextField: TextFieldProps;
  // MuiToggleButton: ToggleButtonProps;
  // MuiToggleButtonGroup: ToggleButtonGroupProps;
  // MuiToolbar: ToolbarProps;
  // MuiTooltip: TooltipProps;
  JrTouchRipple: TouchRippleProps;
  // MuiTypography: TypographyProps;
  // MuiUseMediaQuery: UseMediaQueryOptions;
}